--- src/pages/LoginPage/index.svelte
@@ -1,5 +1,0 @@
-<script lang="ts">
-  import LoginPage from './LoginPage.svelte';
-</script>
-
-<LoginPage />
